moves tests + does not fused if aux requested
